🔄 move RootComp to Core and export via Index
