+++ Framework/Index/Core.js
@@ -1,0 +1,5 @@
+/*****************************************************************************************
+ * exposed to end users, export core functions or comps here
+*****************************************************************************************/
+
+export { default as RootComp } from '../Core/RootComp';
