+++ projects/TemplateApp/App.js
@@ -2,7 +2,7 @@
 * load RootComp here to run the app
 *****************************************************************************************/
 import React from 'react';
-import RootComp from '../../Framework/RootComp';
+import { RootComp } from '../../Framework/Index/Core';
 import { screenMaps, DEFAULT_SCREEN } from './User/ScreenMapper';
 import { NEW_USER_DATA } from './User/Schemas';
